Use getFlowedText in line block elements
A line block is essentially a kind of block quote.  A '|' at the
start of a line forces a new line.  A subsequent line starting
with a space is a continuation of the previous line.  Indentation
within a line block creates a nested line block.  When the line block
(or series of nested blocks) is finished, there should be a blank line
separating it from the next element.

Sphinx produces a 'line_block' event for each block (that is, series of
lines, or nested group of lines), and a 'line' event for each line
within that block.  Currently, each of those line events accumulates
text and then uses the formatted version of that text when creating urwid
widgets.  However, in the case of a continuation line, both lines from
the source file appear in the same line event.  This suggests that the
correct action is to use the flowed version of the text we accumulate.
In other words, the current behavior erroneously preserves the formatting
of continuation lines, whereas by spec, it should not.  The switch to
getFlowedText corrects this.

Additionally, because we are ignoring line_block events, it was not
possible to create nested line_blocks.  By treating them in the same
manner as block quotes, where we put the contents inside of an urwid
widget with left-hand padding, the resulting line block groupings
are nested as expected.